backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.1
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.11.0
matplotlib>=3.7.0
scikit-image>=0.21.0
